modify product 1
