+++ src/GUI/Product/AddNewProduct.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="GUI.Produs.AddNewProduct">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="GUI.Product.AddNewProduct">
   <grid id="27dc6" binding="addNewProductPanel" layout-manager="GridBagLayout">
     <constraints>
       <xy x="20" y="20" width="909" height="671"/>
