+++ src/GUI/Product/ModifyProduct.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="GUI.Produs.ModifyProduct">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="GUI.Product.ModifyProduct">
   <grid id="27dc6" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
