+++ src/GUI/Product/ModifyProduct.java
@@ -1,40 +1,43 @@
-package GUI.Produs;
+package GUI.Product;
 
 import Base.DatabaseConnection;
+import GUI.ProductList;
 
 import javax.swing.*;
 import java.awt.event.*;
 import java.sql.*;
 
-public class AddNewProduct extends JFrame{
-    private JPanel addNewProductPanel;
+public class ModifyProduct extends JFrame {
+    private JPanel modificaProdusPanel;
     private JTextField product_code;
     private JTextField product_category;
     private JTextField product_name;
     private JTextField initial_stock;
     private JTextField company_provenience;
-    private JButton add;
+    private JButton modify;
 
-    public AddNewProduct(String titlu){
-        super(titlu);
+    public ModifyProduct(String tittle){
+        super(tittle);
 
         this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.setContentPane(addNewProductPanel);
+        this.setContentPane(modificaProdusPanel);
         this.pack();
 
-        add.addActionListener(new ActionListener() {
+        String ok1 = new ProductList().getOk1();
+        System.out.println("instance ok1 = " + ok1);
+
+        modify.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-
                 try {
-                    PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("INSERT INTO table1 (provenienta_firma, cod_produs, categorie_produs, nume_produs, stoc_initial) VALUES (?, ?, ?, ?, ?)");
+                    PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("INSERT INTO table1 (company_provenience, product_code, product_category, product_name, initial_stock) VALUES (?, ?, ?, ?, ?)");
                     ps.setString(1, company_provenience.getText());
                     ps.setString(2, product_code.getText());
                     ps.setString(3, product_category.getText());
                     ps.setString(4, product_name.getText());
                     ps.setString(5, initial_stock.getText());
                     ps.executeUpdate();
-                    JOptionPane.showMessageDialog(add, "Produs adaugat cu succes!");
+                    JOptionPane.showMessageDialog(modify, "Product modified ok!");
                     setVisible(false);
                     ps.close();
                 } catch (SQLException ee) {
@@ -43,6 +46,7 @@ public class AddNewProduct extends JFrame{
             }
         });
     }
+
     public JTextField getCompany_provenience() { return company_provenience; }
 
     public JTextField getProduct_code(){
